requirements.txt
pytest==5.4.3
selenium==3.141.0
allure-pytest==2.8.16
pytest-xdist==1.34.0